Initial commit after basic development
Python calculator that I am currently developing and trying to improve.
This is more of a way for me to learn the tkinter library than a way to
make a well-functioning calculator. I found a tutorial on youtube, full
credit to NeuralNine for his "Simple GUI Calculator in Python" video.