Started changed the eval_calc function
The evaluate method in python is built in, and it works great, but it is
not very safe. So I am starting to make a couple methods that will
perform the work itself, which shouldn't be too bad. They are not fully
complete yet, but they have been initialized. The calculator is now
non-functional until further notice