Added in a character
Made the movements for the character available, at least across the x
plane. After I get through the YouTube tutorial, I will go back through
and start editing things as I see fit. For example, the rectangle that
is the player character right now will be the tractor.png file that
can now be found in the folder.