Added in the tractor image!!
Managed to, really easily, get the tractor image inserted for the player
icon. It will work the same as the rectangle, and I even used the
rectangle function to draw it in main, but it all works out. Now I just
need to edit out the background, but that will come later.