Ensured everything is working
The game works as intended now; I need to add a limit to the shots.
The game is too easy to break if you just hold down the A button. I need
to make it so that there can only be a set amount of shots on the screen
at a time